+++ packages/core/test/bundling/router/bundle.golden_symbols.json
@@ -1371,6 +1371,9 @@
     "name": "getOrCreateTNode"
   },
   {
+    "name": "getOrCreateViewRefs"
+  },
+  {
     "name": "getOriginalError"
   },
   {
@@ -1438,6 +1441,9 @@
   },
   {
     "name": "getToken"
+  },
+  {
+    "name": "getViewRefs"
   },
   {
     "name": "handleError"
